requirements.txt
flask==3.1.2
flask-cors==6.0.1
python-dateutil==2.8.2
gunicorn==21.2.0